requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==21.6
python-multipart==0.0.9
pytest==8.3.2